move existing monetary input to provisioning where it is used
