+++ packages/webcomponents/src/components/business-forms/business-form/additional-questions/additional-questions.tsx
@@ -50,7 +50,7 @@ export class AdditionalQuestions {
           <legend part={heading2}>Additional Questions</legend>
           <div class="row gy-3">
             <div class="col-12 col-md-6">
-              <form-control-monetary
+              <form-control-monetary-provisioning
                 name="business_revenue"
                 label="What is the business' estimated annual revenue from its primary business activities?"
                 inputHandler={this.inputHandler}
@@ -60,7 +60,7 @@ export class AdditionalQuestions {
               />
             </div>
             <div class="col-12 col-md-6">
-              <form-control-monetary
+              <form-control-monetary-provisioning
                 name="business_payment_volume"
                 label="What is the business' annual credit card & ACH volume anticipated to process?"
                 inputHandler={this.inputHandler}
